pyproject.toml
[project]
name = "ekan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.2",
    "django>=5.2.7",
    "django-allauth>=65.11.2",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "factory-boy>=3.3.3",
    "faker>=37.8.0",
    "pandas>=2.2.3",
    "pillow>=11.3.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
]
